Extract `line_distance` to `core/view.py`
